refactor constants and collection as modules, as suggested (- WIP #2 -)
