+++ piplicenses/collection.py
